medCONNECT/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
supabase==2.0.3
postgrest==0.13.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Redis
redis==5.0.1
hiredis==2.2.3

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# File Handling
python-magic==0.4.27
Pillow==10.1.0
PyPDF2==3.0.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# AI/ML (for AI services)
openai==1.3.7
langchain==0.0.350

# Utilities
python-dateutil==2.8.2
pytz==2023.3
qrcode[pil]==7.4.2
reportlab==4.0.7

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2





